requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
python-jose
pyjwt
bcrypt